Implement chat
Implement chat as example for how to do websocket messages with the
automaton websocket support.

Implements a chat service that keeps an in-memory list of chat messages
and sends a history of that to new users.

A moving chat sidebar allows the users to chat will all currently
connected other users. If the chat window is closed and the user is
mentioned, show a popover to notify the user.

Greets new users after 3 seconds with a "Welcome @user" message.