Best Pratice: Clone before editing
We have recently come to the conclusion that it is better to clone the
observable object before passing it into the form and then turn off form
isolation for instant write-throughs.

This way the isolation is fundamental and doesn't just disappear when
the form thinks we're done with it.

It also cooperates much better and more natural with browser navigation.

So basically always do it like that in an automaton context.