Upgrade myapp-startup.js
With process v2 structure, we also changed the way we have webpack
prepare our bundles. We now load all processes dynamically and have a
bundle for each.

Changes the startup setup for process v2

Moves the Hub init to the end of the promise chain. This seems to be the
better order to initialize things.