Upgrade example process
Upgrades the example process to the new process v2 structure.

It merges the view component and the transitions of a view state into a
single ViewState object that lives in

processes/<myprocess>/states/<State>.js

Where `myprocess` is the name of the process and `State` the name of the
view state.

The main module now only provides the process scope implementation and a
much slimmer initProcess function.

    export function initProcess(process, scope) {
        return CRUDList;
    }

It now just returns the start state. In general we reference all view
states by their imported ViewState instance now and not by name. Same
for the "to" property in transition map and assignments to the target
property of the transition object.

This new syntax allows to navigate the process via these identifiers
using control-click. The main module also was getting very very long and
the opposite of readable before this.